move numpy, image processing functions out of cli and into respective functions
