+++ src/holo/infra/util/paths.py
@@ -4,7 +4,22 @@ Styled after https://github.com/showyourwork/showyourwork/tree/main.
 
 """
 
+import logging
 from pathlib import Path
+
+logger = logging.getLogger(__name__)
+
+
+def path_check(**kwargs: Path):
+    """Ensure all paths passed in exist."""
+    logger = logging.getLogger(__name__)
+    for variable, path_to_check in kwargs:
+        try:
+            assert isinstance(path_to_check, Path), f"{path_to_check} is not a path."
+            _ = path_to_check.exists()
+        except Exception as e:
+            logger.error(f"{variable} not found/processed at {path_to_check}")
+            raise e
 
 
 # ruff: noqa: D103
